refactor: fixing update function term

+++ backend/src/main/java/com/LAT/backend/repository/TermRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TermRepository extends CrudRepository<Term, Integer> {
+public interface TermRepository extends CrudRepository<Term, Long> {
     Optional<Term> findByDefinition(String name);
 }
